Find dependent files from the .fls generated file
Running the LaTeX compiler with the -recorder option generated a .fls
file which contains the list of all input files (.tex, .sty,...) and
output files (.aux in particular). Then, we can extract the tree of .tex
files. From the .aux files, we get all the .bib files.

The results provided are of course always a bit outdated as they do not
take into account the changes since the last compilation but it will
undoubtedly help with files including using personal macros.